+++ library/custom-shop.php
@@ -96,3 +96,43 @@ if ( ! function_exists( 'bp_shop_page_register_customizer' ) ) {
 				'type'     => 'textarea'))); 
 	}
 }
+
+
+/**
+ * Function to add a row of categories beneath the featured item for quick navigation.
+ * Code adapted from https://stackoverflow.com/questions/21009516/get-woocommerce-product-categories-from-wordpress
+ * 
+ * @since ButtonTheme 1.3.1
+ */
+if ( ! function_exists( 'bp_shop_category_row' ) ) {
+
+	add_action( 'bp_shop_after_featuredcat', 'bp_shop_category_row' );
+	function bp_shop_category_row( $wp_customize ) {
+		$taxonomy     = 'product_cat';
+		$orderby      = 'name';  
+		$show_count   = 0;      // 1 for yes, 0 for no
+		$pad_counts   = 0;      // 1 for yes, 0 for no
+		$hierarchical = 1;      // 1 for yes, 0 for no  
+		$title        = '';  
+		$empty        = 0;
+
+		$args = array(
+				'taxonomy'     => $taxonomy,
+				'orderby'      => $orderby,
+				'show_count'   => $show_count,
+				'pad_counts'   => $pad_counts,
+				'hierarchical' => $hierarchical,
+				'title_li'     => $title,
+				'hide_empty'   => $empty
+		);
+		$all_categories = get_categories( $args );
+		echo "<section class='bp-shop-quicknav'><ul class='menu vertical large-horizontal expanded align-center'>";
+		foreach ($all_categories as $cat) {
+			if($cat->category_parent == 0) {
+				$category_id = $cat->term_id;       
+				echo '<li><a href="'. get_term_link($cat->slug, 'product_cat') .'" class="button small hollow">'. $cat->name .'</a></li>';
+			}       
+		}
+		echo "</ul></section>";
+	}
+}